requirements.txt
google-search-results
pyyaml 